Updates for Test Deployment

+++ Scripts/RemoteManagement/Remove-NAVRemoteWebInstance.ps1
@@ -14,9 +14,15 @@
                 Write-Verbose "Import Module from $($SetupParameters.navServicePath)..."
                 Load-InstanceAdminTools -SetupParameters $SetupParameters
                 Write-Host "Removing Web Client Site for ${ServerInstance}..."
-                Get-NAVWebServerInstance -WebServerInstance $ServerInstance | Remove-NAVWebServerInstance -Force
-                Get-NAVWebServerInstance -WebServerInstance "${ServerInstance}365" | Remove-NAVWebServerInstance -Force
-                Get-NAVWebServerInstance -WebServerInstance "365${ServerInstance}" | Remove-NAVWebServerInstance -Force
+                if ([int]$SetupParameters.mainVersion -ge 110) {
+                    if (Get-NAVWebServerInstance -WebServerInstance $ServerInstance) { Remove-NAVWebServerInstance -WebServerInstance $ServerInstance }
+                    if (Get-NAVWebServerInstance -WebServerInstance "${ServerInstance}365") { Remove-NAVWebServerInstance -WebServerInstance "${ServerInstance}365" }
+                    if (Get-NAVWebServerInstance -WebServerInstance "${ServerInstance}Test") { Remove-NAVWebServerInstance -WebServerInstance "${ServerInstance}Test" }
+                } else {
+                    Get-NAVWebServerInstance -WebServerInstance $ServerInstance | Remove-NAVWebServerInstance -Force
+                    Get-NAVWebServerInstance -WebServerInstance "${ServerInstance}365" | Remove-NAVWebServerInstance -Force
+                    Get-NAVWebServerInstance -WebServerInstance "${ServerInstance}Test" | Remove-NAVWebServerInstance -Force
+                }
                 UnLoad-InstanceAdminTools
                 Start-Sleep -Seconds 2
             } -ArgumentList $SelectedInstance.ServerInstance
